1.流程跑通

+++ DMT.Core.Models/Device/BaseCommand.cs
@@ -294,7 +294,7 @@ namespace DMT.Core.Models
         }
 
 
-        public virtual bool ParseResponse(string content, ref double value)
+        public virtual bool ParseResponse(string content, ref float value)
         {
             return true;
         }
