+++ TAI.Device.Analog/BeamexMC6/BeamexMC6.cs
@@ -17,7 +17,7 @@ namespace TAI.Device
             throw new NotImplementedException();
         }
 
-        public bool GetValue( ChannelType channelType, ref double value)
+        public bool GetValue( ChannelType channelType, ref float value)
         {
             throw new NotImplementedException();
         }
@@ -34,7 +34,7 @@ namespace TAI.Device
             throw new NotImplementedException();
         }
 
-        public bool SetValue( ChannelType channelType, double value)
+        public bool SetValue( ChannelType channelType, float value)
         {
             throw new NotImplementedException();
         }
