+++ TAI.Device.Analog/Fluke/Fluke8846/Commands/GetValue.cs
@@ -25,7 +25,7 @@ namespace TAI.Device.Fluke.D8846
         }
 
 
-        public override bool ParseResponse(string content, ref double value)
+        public override bool ParseResponse(string content, ref float value)
         {
             return true;
         }
